fastapi-benchmark/requirements.txt
fastapi==0.71.0
uvicorn[standard]==0.16.0
python-dotenv==0.19.2
requests==2.27.1
numpy==1.22.1
datasets==1.18.2
scikit-learn==1.0.2
google-api-python-client==2.36.0